Introduce Hotkey Configurator UI and integrate into the main menu
Creates the visual Hotkey Configurator dialog (Options -> Hotkeys
Configuration) allowing users to view, assign, and unbind application
shortcuts. Integrates the hotkey resolution into the input filtering
pipeline and registers Editor and Viewer defaults, safely keeping
direct ProcessKey fallsbacks intact for unmanaged unit test runs.

+++ viewer_view.go
@@ -467,7 +467,7 @@ func (vv *ViewerView) ProcessKey(e *vtinput.InputEvent) bool {
 	case vtinput.VK_F4:
 		vv.HexMode = !vv.HexMode
 		if vv.HexMode {
-			vv.TopOffset &= ^0xF
+			vv.TopOffset &= ^int64(0xF)
 		}
 		return true
 	case vtinput.VK_F6:
